RN ChangeProductStatus implementando no task e testado

+++ src/main/java/com/leilaodequadrinhos/api/model/dao/ProductDao.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface ProductDao extends DAO {
 
 	List<Product> findAllByUser(Long UserId);
-	void changeStatusProduct(Integer id, ProductStatus productStatus);
+	void changeStatusProduct(Long id, ProductStatus productStatus);
 }
